updated and reorganized backend

+++ backend/base/models.py
@@ -9,7 +9,7 @@ class Product(models.Model):
     brand = models.CharField(max_length=200, null=True, blank=True)
     category = models.CharField(max_length=200, null=True, blank=True)
     description = models.TextField(null=True, blank=True)
-    rating = models.IntegerField(null=True, blank=True, default=0)
+    rating = models.DecimalField(null=True, blank=True, default=0, max_digits=7, decimal_places=2)
     numReviews = models.IntegerField(null=True, blank=True, default=0)
     price = models.DecimalField(max_digits=7, decimal_places=2, null=True, blank=True)
     countInStock = models.IntegerField(null=True, blank=True, default=0)
@@ -25,7 +25,7 @@ class Review(models.Model):
     product = models.ForeignKey(Product, on_delete=models.SET_NULL, null=True)
     user = models.ForeignKey(User, on_delete=models.SET_NULL, null=True)
     name = models.CharField(max_length=200, null=True, blank=True)
-    rating = models.IntegerField(null=True, blank=True, default=0)
+    rating = models.DecimalField(null=True, blank=True, default=0, max_digits=7, decimal_places=2)
     comment = models.TextField(null=True, blank=True)
     _id = models.AutoField(primary_key=True, editable=False)
